// frontend/src/types/bindings/Team.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Represents a team in the Bible Challenge.
 */
export type Team = { team_name: string, score: number, buzz_lock_owned: boolean, has_buzzed: boolean, last_buzz_attempt: string | null, };